fix some warnings
git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@4709 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/filters/parser/BaseSplitter/BaseSplitterFileEx.h
@@ -371,7 +371,7 @@ public:
 		BYTE profile, level;
 		unsigned int width, height;
 		unsigned int views;
-		UINT64 crop_left, crop_right, crop_top, crop_bottom;
+		unsigned int crop_left, crop_right, crop_top, crop_bottom;
 		__int64 AvgTimePerFrame;
 
 		struct sar {
